Raffy/requirements.txt
numpy>=1.16.0
scipy>=1.4.1
mir-flare
ase==3.20.1
ray==1.0.1.post1
scikit-learn==0.23.2
